+++ src/credential/status/types.ts
@@ -46,7 +46,6 @@ export const ParsedStatusAttestation = z.object({
  * Shape from parsing a status attestation response in case of error.
  */
 export const InvalidStatusAttestationResponse = z.object({
-  error_description: z.string(),
   error: z.string(),
 });
